change format for sendmessage
